+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/operand/Operator.kt
@@ -1,9 +1,9 @@
-package edu.nextstep.camp.calculator.domain
+package edu.nextstep.camp.calculator.domain.operand
 
 enum class Operator(
     val code: String,
     val operateFun: (Double, Double) -> Double,
-) {
+) : Operand {
     PLUS("+", { a, b -> a + b }),
     MINUS("-", { a, b -> a - b }),
     MULTIPLY("*", { a, b -> a * b }),
@@ -14,6 +14,10 @@ enum class Operator(
             throw IllegalArgumentException("Can not DIVISION Zero")
         }
         return operateFun(first, second)
+    }
+
+    override fun toString(): String {
+        return this.code
     }
 
     companion object {
