+++ domain/src/test/java/edu/nextstep/camp/calculator/domain/OperatorTest.kt
@@ -2,6 +2,7 @@ package edu.nextstep.camp.calculator.domain
 
 import com.google.common.truth.Expect
 import com.google.common.truth.Truth.assertThat
+import edu.nextstep.camp.calculator.domain.operand.Operator
 import org.junit.Rule
 import org.junit.Test
